requirements.txt
argh==0.26.1
dropbox==2.2.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pathtools==0.1.2
pyslack==0.1.3
PyYAML==3.11
redis==2.10.3
requests==2.7.0
urllib3==1.10.4
watchdog==0.8.3
Werkzeug==0.10.4
